% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_file.R
\name{copy_file}
\alias{copy_file}
\title{Copies file from source directory to the proper year directory}
\usage{
copy_file(file)
}
\arguments{
\item{file}{character vector of length one having the basename of the file.}
}
\description{
Copies files from the unprocessed file directory in the the proper CAMP
directory based on date of surgery.
}
