% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_date.R
\name{extract_date}
\alias{extract_date}
\title{Extract date from full path file name}
\usage{
extract_date(path_name)
}
\arguments{
\item{path_name}{character vector of length one having the path and basename
of the file.}
}
\value{
Character vector of length 1 having the date in \%m-\%d-\%Y format
extracted from the file name.
}
\description{
Extract date from full path file name
}
