% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_id.R
\name{extract_id}
\alias{extract_id}
\title{Extract animal ID from full path file name}
\usage{
extract_id(path_name)
}
\arguments{
\item{path_name}{character vector of length one having the path and basename
of the file.}
}
\value{
Character vector of length 1 having the date in \%m-\%d-\%Y format
extracted from the file name.
}
\description{
Uses several huristic techniques based on historical data to isolate the
animal ID.
}
