% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_directory.R
\name{get_directory}
\alias{get_directory}
\title{Get name of directory base on directory type}
\usage{
get_directory(dir_type)
}
\arguments{
\item{dir_type}{character vector of length one having the value of the index
to a specific directory type. Available types are "BASE", "CAMP", and
"UNPROCESSED".}
}
\description{
Get name of directory base on directory type
}
