% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_sequence_char.R
\name{get_next_sequence_char}
\alias{get_next_sequence_char}
\title{Get next sequence character}
\usage{
get_next_sequence_char(path_name)
}
\arguments{
\item{path_name}{character vector of length one having the path and basename
of the file.}
}
\value{
Next sequence character for making ordered filenames when one animal
has more than one surgery sheet per day. The first character returned is
the letter 'A' if there is not a filename for that animal on that date with
the letter A in sequence position.
}
\description{
Get next sequence character
}
