% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sheet_name.R
\name{get_sheet_name}
\alias{get_sheet_name}
\title{Make surgery sheet file name from initial from full path file name}
\usage{
get_sheet_name(path_name)
}
\arguments{
\item{path_name}{character vector of length one having the path and basename
of the file.}
}
\value{
Character vector of length one having properly sequence file name.
}
\description{
Make surgery sheet file name from initial from full path file name
}
