% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_animal.R
\name{is_animal}
\alias{is_animal}
\title{Is ID an animal ID}
\usage{
is_animal(conn, id)
}
\arguments{
\item{conn}{database connection object}

\item{id}{character vector of length one having the ID of the animal}
}
\value{
TRUE if the \code{id} argument represents an animal ID.
}
\description{
Is ID an animal ID
}
