% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_procedure_date.R
\name{is_procedure_date}
\alias{is_procedure_date}
\title{Is the date a procedure date for the specified animal}
\usage{
is_procedure_date(conn, id, date_str)
}
\arguments{
\item{conn}{database connection object}

\item{id}{character vector of length one having the ID of the animal}

\item{date_str}{characer vector of length one having the date of the surgery.}
}
\value{
logical TRUE is the date provided corresponds to a procedure
for the animal indicated.
}
\description{
Uses a simplisted database query to see if there is any type of clinical,
maintenance or research procedure of the indicated date.
}
\details{
This should be updated to look for surgeries when possible.
}
