% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_duplicates.R
\name{list_duplicates}
\alias{list_duplicates}
\title{List any possible file duplicates based on animal name and date.}
\usage{
list_duplicates(all_files)
}
\arguments{
\item{all_files}{character vector with file names of surgery sheets.}
}
\description{
This is only a list of possible duplicates. It is possible to have more than
one PDF generated as a surgery sheet per data for an animal, but it is
rare enough to have someone ensure the contents of the PDFs are infact
different.
}
