% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make new sequenced filename from full path name}
\usage{
make_filename(path_name, sequence_str)
}
\arguments{
\item{path_name}{character vector of length one having the path and basename
of the file.}

\item{sequence_str}{character vector of length one having the characters
that indicate the chronological order of this particular file among the
other files for the same animal on the same date.}
}
\value{
Character vector of length one having the properly sequenced
file name.
}
\description{
Make new sequenced filename from full path name
}
