% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename_A.R
\name{make_filename_A}
\alias{make_filename_A}
\title{Make initial sequenced filename from full path name}
\usage{
make_filename_A(path_name)
}
\arguments{
\item{path_name}{character vector of length one having the path and basename
of the file.}
}
\value{
Character vector of length one having the first sequenced
file name, which contains the letter 'A' as the sequence position indicator.
}
\description{
Make initial sequenced filename from full path name
}
