% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filenames.R
\name{make_filenames}
\alias{make_filenames}
\title{Make new sequenced filenames from full path names}
\usage{
make_filenames(file_list)
}
\arguments{
\item{file_list}{character vector of one or more having the path and basename
of a list of files to be renamed.}
}
\value{
Character vector of properly sequenced file names.
}
\description{
Make new sequenced filenames from full path names
}
