% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_file.R
\name{move_file}
\alias{move_file}
\title{Move file}
\usage{
move_file(file)
}
\arguments{
\item{file}{character vector of length one having the basename of the file.}
}
\description{
Extracts the date from the file name to use in identifying which year
directory to move the file to. The file is then moved and any errors that
ensue are reported.
}
