% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_and_move_files.R
\name{rename_and_move_files}
\alias{rename_and_move_files}
\title{Rename and move files}
\usage{
rename_and_move_files(conn, all_files)
}
\arguments{
\item{conn}{database connection object}

\item{all_files}{character vector of one or more files}
}
\value{
A list containing \code{moved_files}, \code{bad_file_count},
\code{surgery_sheet_count}, and the \code{failed_to_copy_count}.
}
\description{
Takes a list of surgery sheet file names, looks to see if there is an
animal procedure on the date corresponding to the date indicated by the
filename. If there is the file is copied to the appropriate year directory
and renamed.
}
