% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_report.R
\name{rename_report}
\alias{rename_report}
\title{Date stamp report file.}
\usage{
rename_report(path = "")
}
\arguments{
\item{path}{character vector of length one containing the path to add to the
new file name.}
}
\value{
Character vector of length one having the new date and time
stamped file name.
}
\description{
Looks for a report file with the name of \code{renameSurgerySheets.pdf}
and renames it with a call to rmsutilityr::get_dated_filename()
to a date stamped version of \code{Surgery Sheet Report.pdf}.
}
