% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_if_no_files_mounted.R
\name{skip_if_no_files_mounted}
\alias{skip_if_no_files_mounted}
\title{Returns FALSE if database connection is not avaiable.}
\usage{
skip_if_no_files_mounted()
}
\description{
Currently this is a stub
}
